Swap in new ELBO estimator.
