+++ .gitignore
@@ -10,5 +10,4 @@ src/dists.js
 src/inference/enumerate.js
 src/inference/elbo.js
 src/inference/eubo.js
-src/inference/elbo2.js
 src/aggregation/ScoreAggregator.js
