+++ src/inference/elbo.ad.js
@@ -1,20 +1,8 @@
-// Estimates the gradient of the ELBO.
-
-// The estimator used is a combination of the
-// likelihood-ratio/REINFORCE estimator and the "reparameterization
-// trick" based estimator.
-
-// Note that not all parameters are passed explicitly. Parameters are
-// created lazily, and the guide program specifies how they should be
-// initialized.
-
-// Only the gradients of parameters seen during sampling are returned.
-// All other gradients are taken to be zero.
-
 'use strict';
 
 var _ = require('underscore');
 var assert = require('assert');
+var fs = require('fs');
 var util = require('../util');
 var ad = require('../ad');
 var paramStruct = require('../paramStruct');
@@ -24,7 +12,17 @@ module.exports = function(env) {
 
   function ELBO(wpplFn, s, a, options, state, params, step, cont) {
     this.opts = util.mergeDefaults(options, {
-      samples: 1
+      samples: 1,
+      avgBaselines: true,
+      avgBaselineDecay: 0.9,
+      // Weight all factors in the LR term by log p/q.
+      naiveLR: false,
+      // Write a DOT file representation of first graph to disk.
+      dumpGraph: false,
+      // Use local weight of one for all sample and factor nodes. This
+      // is useful in combination with the dumpGraph option for
+      // understanding/debugging weight propagation.
+      debugWeights: false
     });
 
     // The current values of all initialized parameters.
@@ -32,17 +30,177 @@ module.exports = function(env) {
     this.params = params;
 
     this.step = step;
+    this.state = state;
     this.cont = cont;
 
     this.wpplFn = wpplFn;
     this.s = s;
     this.a = a;
 
-    this.mapDataState = {};
+    // Initialize mapData state.
+    this.mapDataStack = [{multiplier: 1}];
+    this.mapDataIx = {};
+
+    if (!_.has(this.state, 'baselines')) {
+      this.state.baselines = {};
+    }
+    this.baselineUpdates = {};
 
     this.coroutine = env.coroutine;
     env.coroutine = this;
   }
+
+  function top(stack) {
+    return stack[stack.length - 1];
+  }
+
+  // The strategy taken here is to build a graph to (coarsely) track
+  // dependency information which we use for variance reduction. This
+  // simple approach builds a graph that represents:
+
+  // 1. The order in which random choices are made.
+  // 2. The conditional independence information from mapData.
+
+  // This is used to remove some unnecessary (i.e. upstream) terms
+  // from the weighting applied to each "grad logq" factor in the LR
+  // part of the objective. This improves on the naive implementation
+  // which weights each factor by logq - logp of the full execution.
+
+  // The graph is built as the program executes. A separate pass then
+  // propagates weights back up the graph, taking account of any
+  // sub-sampling of data that happens at mapData. After this pass, a
+  // node's weight property is the correct weighting for the
+  // corresponding grad logq factor of the objective.
+
+  var nodeid = 0;
+
+  function RootNode() {
+    this.id = nodeid++;
+    this.parents = [];
+    this.weight = 0;
+  }
+
+  function SampleNode(parent, logp, logq, reparam, address, targetDist, guideDist, value, multiplier, debug) {
+    this.id = nodeid++;
+    var _logp = ad.value(logp);
+    var _logq = ad.value(logq);
+    this.parents = [parent];
+    this.logp = logp;
+    this.logq = logq;
+    this.weight = debug ? 1 : _logq - _logp;
+    this.reparam = reparam;
+    this.address = address;
+    this.multiplier = multiplier;
+    // Debug info.
+    this.targetDist = targetDist;
+  }
+
+  SampleNode.prototype.label = function() {
+    return [
+      this.targetDist.meta.name + '(' + this.id + ')',
+      'w=' + this.weight,
+      'm=' + this.multiplier
+    ].join('\\n');
+  };
+
+  function FactorNode(parent, score, multiplier, debug) {
+    this.id = nodeid++;
+    var _score = ad.value(score);
+    this.parents = [parent];
+    this.score = score;
+    this.weight = debug ? 1 : -_score;
+    this.multiplier = multiplier;
+  }
+
+  FactorNode.prototype.label = function() {
+    return [
+      'Factor' + '(' + this.id + ')',
+      'w=' + this.weight,
+      'm=' + this.multiplier
+    ].join('\\n');
+  };
+
+  // Created when entering mapData.
+  function SplitNode(parent, batchSize, n, joinNode) {
+    this.id = nodeid++;
+    this.parents = [parent];
+    this.batchSize = batchSize;
+    this.n = n; // data.length
+    this.joinNode = joinNode;
+    this.weight = 0;
+  }
+
+  // Created when leaving mapData.
+  function JoinNode() {
+    this.id = nodeid++;
+    this.parents = [];
+    this.weight = 0;
+  }
+
+  function propagateWeights(nodes) {
+    // Note that this modifies the weights of the graph in-place.
+    var i = nodes.length;
+    while(--i) {
+      var node = nodes[i];
+      if (node instanceof SplitNode) {
+        // Account for (a) the fact that we (potentially) only looked
+        // at a subset of the data (i.e. used mini-batches) and (b)
+        // the weights downstream of the associated join node will
+        // have been included in the split node's weight once for each
+        // execution of the observation function.
+        node.weight = (node.n / node.batchSize) * node.weight - ((node.n - 1) * node.joinNode.weight);
+      }
+      node.parents.forEach(function(parent) {
+        parent.weight += node.weight;
+      });
+    }
+  };
+
+  var edge = function(parent, child) {
+    return '  ' + parent.id + ' -> ' + child.id + ';';
+  };
+
+  var shape = function(node, shape) {
+    return '  ' + node.id + ' [shape = "' + shape + '"]';
+  };
+
+  var label = function(node) {
+    return '  ' + node.id + ' [label = "' + node.label() + '"]';
+  };
+
+  function generateDot(nodes) {
+    var edges = [];
+    var append = function(x) { edges.push(x); };
+    nodes.forEach(function(node) {
+      if (node instanceof FactorNode) {
+        append(shape(node, 'box'));
+      }
+      if (node instanceof RootNode ||
+          node instanceof JoinNode ||
+          node instanceof SplitNode) {
+        append(shape(node, 'point'));
+      }
+      if (node.label) {
+        append(label(node));
+      }
+      node.parents.forEach(function(parent) {
+        append(edge(parent, node));
+      });
+    });
+    return 'digraph {\n' + edges.join('\n') + '\n}\n';
+  };
+
+  function checkScoreIsFinite(score, source) {
+    var _score = ad.value(score);
+    if (!isFinite(_score)) { // Also catches NaN.
+      var msg = 'ELBO: The score of the previous sample under the ' +
+            source + ' program was ' + _score + '.';
+      if (_.isNaN(_score)) {
+        msg += ' Reducing the step size may help.';
+      }
+      throw new Error(msg);
+    }
+  };
 
   ELBO.prototype = {
 
@@ -68,6 +226,7 @@ module.exports = function(env) {
         function() {
           paramStruct.divEq(grad, this.opts.samples);
           elbo /= this.opts.samples;
+          this.updateBaselines();
           env.coroutine = this.coroutine;
           return this.cont(grad, elbo);
         }.bind(this));
@@ -77,74 +236,126 @@ module.exports = function(env) {
     // Compute a single sample estimate of the gradient.
 
     estimateGradient: function(cont) {
-      'use ad';
-
       // paramsSeen tracks the AD nodes of all parameters seen during
       // a single execution. These are the parameters for which
       // gradients will be computed.
       this.paramsSeen = {};
-      this.logp = this.logq = this.logr = 0;
+
+      // This tracks nodes as we encounter them which saves doing a
+      // topological sort later on.
+      this.nodes = [];
+
+      var root = new RootNode();
+      this.prevNode = root; // prevNode becomes the parent of the next node.
+      this.nodes.push(root);
 
       return this.wpplFn(_.clone(this.s), function() {
 
-        var _logq = ad.value(this.logq);
-        var _logp = ad.value(this.logp);
-        checkScoreIsFinite(_logq, 'guide');
-        checkScoreIsFinite(_logp, 'target');
+        propagateWeights(this.nodes);
 
-        var scoreDiff = _logq - _logp;
-        assert.ok(typeof scoreDiff === 'number');
-
-        // Objective.
-
-        // We could use the hybrid objective in all situations, but
-        // for statistical efficiency we drop terms with zero
-        // expectation where possible.
-
-        var useLR = sameAdNode(this.logq, this.logr);
-
-        // Sanity check.
-
-        if (useLR) {
-          // log p isn't expected to depend on the parameters unless
-          // we use reparameterization.
-          assert.ok(typeof this.logp === 'number');
+        if (this.step === 0 && this.iter === 0 && this.opts.dumpGraph) {
+          // To vizualize with Graphviz use:
+          // dot -Tpng -O deps.dot
+          var dot = generateDot(this.nodes);
+          fs.writeFileSync('deps.dot', dot);
         }
 
-        // The objective used could change across steps, but for
-        // simplicity only report on the first step.
+        var ret = this.buildObjective();
 
-        if (this.opts.verbose && this.iter === 0 && this.step === 0) {
-          // Here PW stands for "path-wise" estimator, aka the reparam
-          // trick.
-          var estName =
-                useLR ? 'LR' :
-                (typeof this.logr === 'number') ? 'PW' :
-                'hybrid';
-
-          console.log('ELBO: Using ' + estName + ' estimator.');
-        }
-
-        var objective = useLR ?
-              this.logq * scoreDiff :
-              this.logr * scoreDiff + this.logq - this.logp;
-
-        if (ad.isLifted(objective)) { // handle guides with zero parameters
-          objective.backprop();
+        if (ad.isLifted(ret.objective)) { // Handle programs with zero random choices.
+          ret.objective.backprop();
         }
 
         var grads = _.mapObject(this.paramsSeen, function(params) {
           return params.map(ad.derivative);
         });
 
-        return cont(grads, -scoreDiff);
+        return cont(grads, ret.elbo);
 
       }.bind(this), this.a);
 
     },
 
+    buildObjective: function() {
+      'use ad';
+      var naiveLR = this.opts.naiveLR;
+      var rootNode = this.nodes[0];
+      assert.ok(rootNode instanceof RootNode);
+      assert.ok(_.isNumber(rootNode.weight));
+
+      var objective = this.nodes.reduce(function(acc, node) {
+        if (node instanceof SampleNode && node.reparam) {
+          return acc + node.multiplier * (node.logq - node.logp);
+        } else if (node instanceof SampleNode) {
+          assert.ok(!node.param);
+          var weight = naiveLR ? rootNode.weight : node.weight;
+          assert.ok(_.isNumber(weight));
+          var b = this.computeBaseline(node.address, weight);
+          return acc + node.multiplier * ((node.logq * (weight - b)) - node.logp);
+        } else if (node instanceof FactorNode) {
+          return acc - node.multiplier * node.score;
+        } else {
+          return acc;
+        }
+      }.bind(this), 0);
+      var elbo = -rootNode.weight;
+      return {objective: objective, elbo: elbo};
+    },
+
+    computeBaseline: function(address, weight) {
+      if (!this.opts.avgBaselines) {
+        return 0;
+      }
+
+      var baselines = this.state.baselines;
+      var baselineUpdates = this.baselineUpdates;
+
+      // Accumulate the mean of the weights for each factor across
+      // all samples taken this step. These are incorporated into
+      // the running average once all samples have been taken.
+      // Note that each factor is not necessarily encountered the
+      // same number of times.
+
+      if (!_.has(baselineUpdates, address)) {
+        baselineUpdates[address] = {n: 1, mean: weight};
+      } else {
+        var prev = baselineUpdates[address];
+        var n = prev.n + 1;
+        var mean = (prev.n * prev.mean + weight) / n;
+        baselineUpdates[address].n = n;
+        baselineUpdates[address].mean = mean;
+      }
+
+      // During the first step we'd like to use the weight as the
+      // baseline. The hope is that this strategy might avoid very
+      // large gradients on the first step. If the initial baseline
+      // was zero, these large gradients may cause optimization
+      // methods with adaptive step sizes to reduce the step size (for
+      // associated parameters) more than will be necessary once the
+      // baseline takes effect. This might slow the initial phase of
+      // optimization. However, using exactly the weight would cause
+      // the gradient to be zero which in turn would trigger a warning
+      // from Optimize. To avoid this we scale the weight and use that
+      // as the initial baseline.
+
+      return _.has(baselines, address) ? baselines[address] : weight * .99;
+    },
+
+    updateBaselines: function() {
+      var decay = this.opts.avgBaselineDecay;
+      var baselines = this.state.baselines;
+      // Note that this leaves untouched the estimate of the average
+      // weight for any factors not seen during this step.
+      _.each(this.baselineUpdates, function(obj, address) {
+        baselines[address] = _.has(baselines, address) ?
+          decay * baselines[address] + (1 - decay) * obj.mean :
+          obj.mean;
+      }, this);
+    },
+
     sample: function(s, k, a, dist, options) {
       options = options || {};
+
       var guideDist;
       if (options.guide) {
         guideDist = options.guide;
@@ -157,132 +368,131 @@ module.exports = function(env) {
           console.log('ELBO: Defaulting to mean-field for one or more choices.');
         }
       }
-      var guideVal = this.sampleGuide(guideDist, options);
-      this.sampleTarget(dist, guideVal);
-      return k(s, guideVal);
+
+      var ret = this.sampleGuide(guideDist, options);
+      var val = ret.val;
+
+      var logp = dist.score(val);
+      var logq = guideDist.score(val);
+      checkScoreIsFinite(logp, 'target');
+      checkScoreIsFinite(logq, 'guide');
+
+      var m = top(this.mapDataStack).multiplier;
+
+      var node = new SampleNode(
+        this.prevNode, logp, logq,
+        ret.reparam, a, dist, guideDist, val, m, this.opts.debugWeights);
+
+      this.prevNode = node;
+      this.nodes.push(node);
+
+      return k(s, val);
     },
 
     sampleGuide: function(dist, options) {
-      'use ad';
-
-      var val;
+      var val, reparam;
 
       if ((!_.has(options, 'reparam') || options.reparam) &&
           dist.base && dist.transform) {
         // Use the reparameterization trick.
-
         var baseDist = dist.base();
         var z = baseDist.sample();
-        this.logr += baseDist.score(z);
         val = dist.transform(z);
-        this.logq += dist.score(val);
-
+        reparam = true;
       } else if (options.reparam && !(dist.base && dist.transform)) {
-        // Warn when reparameterization is explicitly requested but
-        // isn't supported by the distribution.
         throw dist + ' does not support reparameterization.';
       } else {
         val = dist.sample();
-        var score = dist.score(val);
-
-        if (sameAdNode(this.logq, this.logr)) {
-          // The reparameterization trick has not been used yet.
-          // Continue representing logq and loqr with the same ad
-          // node.
-          this.logr += score;
-          this.logq = this.logr;
-        } else {
-          // The reparameterization trick has been used earlier in the
-          // execution. Update logq and logr independently.
-          this.logr += score;
-          this.logq += score;
-        }
+        reparam = false;
       }
 
-      return val;
+
+      return {val: val, reparam: reparam};
     },
 
-    sampleTarget: function(dist, guideVal) {
-      'use ad';
-      this.logp += dist.score(guideVal);
-    },
-
-    factor: function(s, k, a, score) {
-      'use ad';
-      this.logp += score;
+    factor: function(s, k, a, score, name) {
+      if (!isFinite(ad.value(score))) {
+        throw new Error('ELBO: factor score is not finite.');
+      }
+      var m = top(this.mapDataStack).multiplier;
+      var node = new FactorNode(
+        this.prevNode, score, m, this.opts.debugWeights);
+      this.prevNode = node;
+      this.nodes.push(node);
       return k(s);
     },
 
     mapDataFetch: function(data, batchSize, address) {
 
+      // Compute batch indices.
+
       var ix;
-      if (batchSize === data.length) {
-        // Use all the data, in order.
-        ix = null;
+      if (_.has(this.mapDataIx, address)) {
+        ix = this.mapDataIx[address];
       } else {
-        ix = _.times(batchSize, function() {
-          return Math.floor(util.random() * data.length);
-        });
+        if (batchSize === data.length) {
+          // Use all the data, in order.
+          ix = null;
+        } else {
+          ix = _.times(batchSize, function() {
+            return Math.floor(util.random() * data.length);
+          });
+        }
+        // Store batch indices so that we can use the same mini-batch
+        // across samples.
+        this.mapDataIx[address] = ix;
       }
 
-      // Store the info needed to compute the correction to account
-      // for the fact we only looked at a subset of the data.
+      if (batchSize > 0) {
+        var joinNode = new JoinNode();
+        var splitNode = new SplitNode(this.prevNode, batchSize, data.length, joinNode);
+        this.nodes.push(splitNode);
 
-      assert.ok(!this.mapDataState[address]);
-      this.mapDataState[address] = {
-        logp: this.logp,
-        logq: this.logq,
-        logr: this.logr,
-        multiplier: batchSize > 0 ? (data.length / batchSize) - 1 : 0
-      };
+        // Compute the multiplier required to account for the fact we're
+        // only looking at a subset of the data.
+        var thisM = data.length / batchSize;
+        var prevM = top(this.mapDataStack).multiplier;
+        var multiplier = thisM * prevM;
+
+        this.mapDataStack.push({
+          splitNode: splitNode,
+          joinNode: joinNode,
+          multiplier: multiplier
+        });
+      } else {
+        // Signal to mapDataFinal that the batch was empty.
+        this.mapDataStack.push(null);
+      }
 
       return ix;
     },
 
-    mapDataFinal: function(address) {
-      'use ad';
-
-      var state = this.mapDataState[address];
-      assert.ok(state !== undefined);
-
-      var noreparam = sameAdNode(this.logq, this.logr);
-      var m = state.multiplier;
-
-      this.logp += m * (this.logp - state.logp);
-      this.logq += m * (this.logq - state.logq);
-      if (noreparam) {
-        // The reparameterization trick has not been used yet.
-        // Continue representing logq and loqr with the same ad node.
-        this.logr = this.logq;
-      } else {
-        this.logr += m * (this.logr - state.logr);
-      }
-
-      this.mapDataState[address] = undefined;
+    mapDataEnter: function() {
+      // For every observation function, set the current node back to
+      // the split node.
+      this.prevNode = top(this.mapDataStack).splitNode;
     },
 
-    incrementalize: env.defaultCoroutine.incrementalize,
-    constructor: ELBO
+    mapDataLeave: function() {
+      // Hook-up the join node to the last node on this branch. If
+      // there were no sample/factor nodes created in the observation
+      // function then this connects the join node directly to the
+      // split node. The correction applied to split nodes in
+      // propagateWeights requires such edges to be present for
+      // correctness.
+      top(this.mapDataStack).joinNode.parents.push(this.prevNode);
+    },
+
+    mapDataFinal: function(address) {
+      var top = this.mapDataStack.pop();
+      if (top !== null) {
+        var joinNode = top.joinNode;
+        this.prevNode = joinNode;
+        this.nodes.push(joinNode);
+      }
+    }
 
   };
-
-  function sameAdNode(a, b) {
-    // We can't use === directly within an ad transformed function as
-    // doing so checks the equality of the values stored at the nodes
-    // rather than the nodes themselves.
-    return a === b;
-  }
-
-  function checkScoreIsFinite(score, source) {
-    if (!_.isFinite(score)) { // Also catches NaN.
-      var msg = 'ELBO: The score of the previous sample under the ' +
-            source + ' program was ' + score + '.';
-      if (_.isNaN(score)) {
-        msg += ' Reducing the step size may help.';
-      }
-      throw new Error(msg);
-    }
-  }
 
   return function() {
     var coroutine = Object.create(ELBO.prototype);
